only_ocr/requirements.txt
python-mecab-ko
streamlit
spacy-streamlit
pororo
fastapi
uvicorn

opencv-python
scikit-image

